Make UsageFormatter abstract, Create DefaultUsageFormatter
